+++ artifex/models/classification/__init__.py
@@ -2,11 +2,12 @@ from .classification_model import ClassificationModel
 from .binary_classification import SpamDetection
 from .multi_class_classification import EmotionDetection, IntentClassifier, SentimentAnalysis, \
     TopicClassification
-from .multi_label_classification import Guardrail
+from .multi_label_classification import UserQueryGuardrail, LLMOutputGuardrail
     
 __all__ = [
     "ClassificationModel",
-    "Guardrail",
+    "UserQueryGuardrail",
+    "LLMOutputGuardrail",
     "EmotionDetection",
     "IntentClassifier",
     "SentimentAnalysis",