+++ artifex/models/classification/multi_label_classification/__init__.py
@@ -1,5 +1,7 @@
-from .guardrail import Guardrail
+from .user_query_guardrail import UserQueryGuardrail
+from .llm_output_guardrail import LLMOutputGuardrail
 
 __all__ = [
-    "Guardrail",
+    "UserQueryGuardrail",
+    "LLMOutputGuardrail",
 ]